feat: #2 add api versioning
- Reorganize folder structure
- Add v1 router and its subrouters
