+++ app/api/v1/__init__.py
@@ -1,0 +1,3 @@
+from .router import router as v1_router
+
+__all__ = ["v1_router"]
