+++ app/api/v1/router.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+from app.api.v1.routers import token_router, user_router
+
+router = APIRouter(prefix="/api/v1", tags=["v1"])
+
+router.include_router(token_router)
+router.include_router(user_router)
