+++ app/dependencies/auth.py
@@ -7,7 +7,7 @@ from app.core.security import decode_token, verify_password
 from app.database.session import DbSession
 from app.exceptions import NotFoundError, UnautorizedError
 from app.models.user import User
-from app.operations import user as crud_user
+from app.operations.user import crud as crud_user
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
